+++ app/src/main/java/eu/kanade/tachiyomi/data/backup/restore/RestoreOptions.kt
@@ -42,7 +42,7 @@ data class RestoreOptions(
                 setter = { options, enabled -> options.copy(appSettings = enabled) },
             ),
             Entry(
-                label = TDMR.strings.pref_novel_extension_repos,
+                label = MR.strings.label_extension_repos,
                 getter = RestoreOptions::extensionRepositories,
                 setter = { options, enabled -> options.copy(extensionRepositories = enabled) },
             ),