+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is a modified version of [Keep a Changelog](https://keepachangelog.co
 ### Added
 
 ### Improved
-- Refresh the plugin list when a JS repository is toggled, and stop one unreachable store from emptying the whole extension fetch [@Rojikku](https://github.com/Rojikku) [#391](https://github.com/tsundoku-otaku/tsundoku/pull/391)
+- Novel Extension Repo > Extension Store + Related fixes [@Rojikku](https://github.com/Rojikku) [#391](https://github.com/tsundoku-otaku/tsundoku/pull/391)
 - Import Stub isNovel improvements, fix last_read backup restore error, change quick + migrate to durable background jobs [@Rojikku](https://github.com/Rojikku) [#392](https://github.com/tsundoku-otaku/tsundoku/pull/392)
 - Handle picture/source/video/audio tags in text view [@mrissaoussama](https://github.com/mrissaoussama) [#381](https://github.com/tsundoku-otaku/tsundoku/pull/381)
 - avoid list copies during backup restore [@mrissaoussama](https://github.com/mrissaoussama) [#389](https://github.com/tsundoku-otaku/tsundoku/pull/389)